requirements.txt


requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
trafilatura>=1.9.0
pypdf>=4.0.0
google-genai>=1.0.0
langchain-text-splitters>=0.2.0
faiss-cpu>=1.8.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
numpy>=1.26.0
python-dotenv>=1.0.0